requirements.txt
langchain
langchain-google-genai
langchain-community
sqlalchemy
streamlit
python-dotenv
psycopg2-binary
pymysql
cryptography
